fixed bugs

+++ main/alyukabond/prixod.py
@@ -34,55 +34,24 @@ def color():
         db.session.delete(color)
         db.session.commit()
         return jsonify(msg="Success")
-    
 
-# @bp.route('/aluminy-thickness', methods=['GET', 'POST', 'DELETE'])
+
+# @bp.route('/add-balance-user', methods=['GET', 'POST'])
 # @jwt_required()
-# def aluminy_thickness():
-#     id = request.args.get('id')
-#     if request.method == 'GET':
-#         if id is not None:
-#             thkn = AluminyThickness.query.get(id)
-#             return jsonify(aluminy_thickness_schema.dump(thkn))
-#         thkns = AluminyThickness.query.all()
-#         return jsonify(aluminy_thickness_schemas.dump(thkns))
-#     elif request.method == 'POST':
-#         data = request.get_json()
-#         thkn = AluminyThickness(thickness=data.get('thickness'))
-#         db.session.add(thkn)
-#         db.session.commit()
-#         return jsonify(msg="Success")
+# def add_balance_user():
+#     if request.method=='GET':
+#         users = AddBalanceUser.query.all()
+#         return jsonify(add_balance_users_schema.dump(users))
 #     else:
-#         thkn = AluminyThickness.query.get(id)
-#         db.session.delete(thkn)
-#         db.session.commit()
-#         return jsonify(msg="Success")
-
-
-# @bp.route('/alyukabond-length', methods=['GET', 'POST', 'DELETE'])
-# @jwt_required()
-# def alyukabond_length():
-#     id = request.args.get('id')
-#     if request.method == 'GET':
-#         if id is not None:
-#             thkn = AlyukabondLength.query.get(id)
-#             return jsonify(alyukabond_length_schema.dump(thkn))
-#         thkns = AlyukabondLength.query.all()
-#         return jsonify(alyukabond_length_schemas.dump(thkns))
-#     elif request.method == 'POST':
 #         data = request.get_json()
-#         thkn = AlyukabondLength(length=data.get('thickness'))
-#         db.session.add(thkn)
-#         db.session.commit()
-#         return jsonify(msg="Success")
-#     else:
-#         thkn = AlyukabondLength.query.get(id)
-#         db.session.delete(thkn)
+#         user = AddBalanceUser(name=data.get('name'))
+#         db.session.add(user)
 #         db.session.commit()
 #         return jsonify(msg="Success")
 
 
 @bp.route('/exchange-rate', methods=['GET', 'POST'])
+@jwt_required()
 def exchange_rate():
     if request.method == 'GET':
         rate = ExchangeRate.query.order_by(ExchangeRate.id.desc()).first()
@@ -128,9 +97,10 @@ def alyuminy_material():
                 provider = data.get("provider")
             )
             db.session.add(nakladnoy)
+            db.session.flush()
             for obj in data.get("aluminy"):
-                add_aluminy_amount(thickness=obj.get('thickness'), width=obj.get('list_width', 1.22),
-                    color_id=obj.get('color_id'), length=obj.get('list_length'), roll_weight=obj.get('roll_weight'), quantity=obj.get('quantity'))
+                # add_aluminy_amount(thickness=obj.get('thickness'), width=obj.get('list_width', 1.22),
+                    # color_id=obj.get('color_id'), length=obj.get('list_length'), roll_weight=obj.get('roll_weight'), quantity=obj.get('quantity'))
                 material = Aluminy(
                     color_id = obj.get('color_id'),
                     nakladnoy_id = nakladnoy.id,
@@ -142,7 +112,8 @@ def alyuminy_material():
                     price = obj.get('price'),
                     quantity = obj.get('quantity'),
                     partiya = nakladnoy.partiya, 
-                    date = nakladnoy.date
+                    date = nakladnoy.date,
+                    provider = nakladnoy.provider
                 )
                 db.session.add(material)
             balance_minus(data.get("payed_price_d"))
@@ -153,31 +124,39 @@ def alyuminy_material():
     elif request.method == 'PUT' or request.method == 'PATCH':
         id = request.args.get('material_id')
         if user.role in ['a']:
-            data = request.get_json()
-            try:
-                material = Aluminy.query.get(id)
-                surface = update_aluminy_amount(material=material, thickness=data.get('thickness', material.thickness), color=data.get('color_id', material.color_id),
-                    list_length=data.get('list_length', material.list_length), list_width=data.get('list_width', material.list_width),
-                    roll_weight=data.get('roll_weight', material.roll_weight), quantity=data.get('quantity', material.quantity))
-                material.color_id = data.get('color_id', material.color_id)
-                material.thickness = data.get('thickness', material.thickness)
-                material.list_width = data.get('list_width', material.list_width)
-                material.list_length = data.get('list_length', material.list_length)
-                material.quantity = data.get('quantity', material.quantity)
-                material.roll_weight = data.get('roll_weight', material.roll_weight)
-                material.price_per_kg = data.get('price_per_kg', material.price_per_kg)
-                db.session.commit()
-                return jsonify(msg='Success')
-            except AssertionError as err:
-                return jsonify(msg=f"{str(err)}"), 400
+            material = Aluminy.query.get(id)
+            if material.editable == True:
+                data = request.get_json()
+                try:
+                    # surface = update_aluminy_amount(material=material, thickness=data.get('thickness', material.thickness), color=data.get('color_id', material.color_id),
+                    #     list_length=data.get('list_length', material.list_length), list_width=data.get('list_width', material.list_width),
+                    #     roll_weight=data.get('roll_weight', material.roll_weight), quantity=data.get('quantity', material.quantity))
+                    material.color_id = data.get('color_id', material.color_id)
+                    material.thickness = data.get('thickness', material.thickness)
+                    material.list_width = data.get('list_width', material.list_width)
+                    material.list_length = data.get('list_length', material.list_length)
+                    material.quantity = data.get('quantity', material.quantity)
+                    material.roll_weight = data.get('roll_weight', material.roll_weight)
+                    material.price_per_kg = data.get('price_per_kg', material.price_per_kg)
+                    db.session.commit()
+                    return jsonify(msg='Success')
+                except AssertionError as err:
+                    return jsonify(msg=f"{str(err)}"), 400
+            return  jsonify(msg="Неотредактируемый объект"), 400
         return jsonify(msg="У вас нет полномочий на это действие"), 401
     elif request.method == 'DELETE':
         if user.role == 'a':
             id = request.args.get("material_id")
-            material = db.get_or_404(Aluminy, id)
-            db.session.delete(material)
-            db.session.commit()
-            return jsonify(msg="Deleted")
+            nakladnoy = db.get_or_404(AluminyNakladnoy, id)
+            if nakladnoy.editable == True:
+                balance_add(nakladnoy.payed_price_d)
+                aluminies = Aluminy.query.filter_by(nakladnoy_id=nakladnoy.id).all()
+                db.session.delete(nakladnoy)
+                for aluminy in aluminies:
+                    db.session.delete(aluminy)
+                    db.session.commit()
+                return jsonify(msg="Deleted")
+            return  jsonify(msg="Неотредактируемый объект"), 400
         return jsonify(msg="У вас нет полномочий на это действие"), 401
 
 
@@ -197,13 +176,16 @@ def glue_material():
             return jsonify(data)
         if material_id is not None:
             return jsonify(glue_schema.dump(Glue.query.get_or_404(material_id)))
-        materials = Glue.query.all()
+        materials = Glue.query.order_by(Glue.date.desc()).all()
         return jsonify(glue_schemas.dump(materials))
     elif request.method == 'POST':
         try:
             data = request.get_json()
-            add_glue_amount(thickness=data.get("thickness"), width=data.get("width", 1.22), length=data.get("length"), roll_weight=data.get("weight"), quantity=data.get("quantity"))
-            material = Glue(**data)
+            # add_glue_amount(thickness=data.get("thickness"), width=data.get("width", 1.22), length=data.get("length"), roll_weight=data.get("weight"), quantity=data.get("quantity"))
+            weight = data['weight']
+            del data['weight']
+            surface = data.get('width') * data.get('length') * data.get('quantity')
+            material = Glue(**data, weight=weight * 1000, surface=surface)
             db.session.add(material)
             balance_minus(data.get("payed_price_d"))
             db.session.commit()
@@ -212,31 +194,45 @@ def glue_material():
                 return jsonify(msg=f"{str(err)}"), 400
     elif request.method == 'PUT' or request.method == 'PATCH':
         if user.role in ['a']:
-            try:
-                data = request.get_json()
-                weight = data.get("weight") * 1000 if data.get("weight") else material.weight
-                del data['weight']
-                material = Glue.query.get(id)
-                update_glue_amount(material=material, thickness=data.get('thickness', material.thickness), length=data.get('length', material.length),
-                     width=data.get('width', material.width), roll_weight=weight, quantity=data.get('quantity', material.quantity))
-                material.thickness = data.get('thickness', material.thickness)
-                material.width = data.get('width', material.width)
-                material.length = data.get('length', material.length)
-                material.quantity = data.get('quantity', material.quantity)
-                material.weight = weight
-                material.price_per_kg = data.get('price_per_kg', material.price_per_kg)
-                material.provider = data.get('provider', material.provider)
-                db.session.commit()
-                return jsonify(msg='Success')
-            except AssertionError as err:
-                return jsonify(msg=f"{str(err)}"), 400
+            material = Glue.query.get(id)
+            if material.editable == True:
+                try:
+                    data = request.get_json()
+                    weight = float(data.get("weight")) * 1000 if data.get("weight") else material.weight
+                    extra_sum = data.get('total_price_d', material.total_price_d) - material.total_price_d
+                    del data['weight']
+                    # update_glue_amount(material=material, thickness=data.get('thickness', material.thickness), length=data.get('length', material.length),
+                    #     width=data.get('width', material.width), roll_weight=weight, quantity=data.get('quantity', material.quantity))
+                    material.thickness = data.get('thickness', material.thickness)
+                    material.width = data.get('width', material.width)
+                    material.length = data.get('length', material.length)
+                    material.quantity = data.get('quantity', material.quantity)
+                    material.weight = weight
+                    material.surface = float(material.width) * float(material.length) * float(material.quantity)
+                    material.price_per_kg = data.get('price_per_kg', material.price_per_kg)
+                    material.total_price_d = data.get('total_price_d', material.total_price_d)
+                    material.total_price_s = data.get('total_price_s', material.total_price_s)
+                    material.payed_price_d = material.payed_price_d
+                    material.payed_price_s = material.payed_price_s
+                    material.debt_d = material.total_price_d - material.payed_price_d
+                    material.debt_s = material.total_price_s - material.payed_price_s
+                    material.provider = data.get('provider', material.provider)
+                    db.session.commit()
+                    balance_add(extra_sum)
+                    return jsonify(msg='Success')
+                except AssertionError as err:
+                    return jsonify(msg=f"{str(err)}"), 400
+            return  jsonify(msg="Неотредактируемый объект"), 400
         return jsonify(msg="У вас нет полномочий на это действие"), 401
     elif request.method == 'DELETE':
         if user.role == 'a':
             material = db.get_or_404(Glue, id)
-            db.session.delete(material)
-            db.session.commit()
-            return jsonify(msg="Deleted")
+            if material.editable == True:
+                balance_add(material.payed_price_d)
+                db.session.delete(material)
+                db.session.commit()
+                return jsonify(msg="Deleted")
+            return  jsonify(msg="Неотредактируемый объект"), 400
         return jsonify(msg="У вас нет полномочий на это действие"), 401
 
 
@@ -255,7 +251,7 @@ def sticker_material():
             return jsonify(data)
         if material_id is not None:
             return jsonify(sticker_schema.dump(Sticker.query.get_or_404(material_id)))
-        materials = Sticker.query.all()
+        materials = Sticker.query.order_by(Sticker.date.desc()).all()
         return jsonify(sticker_schemas.dump(materials))
     elif request.method == 'POST':
         try:
@@ -271,19 +267,21 @@ def sticker_material():
                 provider = data.get("provider")
             )
             db.session.add(nakladnoy)
+            db.session.flush()
             for obj in data.get("sticker"):
-                surface = add_sticker_amount(width=obj.get('width', 1.22),  type=obj.get('type_sticker'), length=obj.get('length'), quantity=obj.get('quantity'))
+                # surface = add_sticker_amount(width=obj.get('width', 1.22),  type=obj.get('type_sticker'), length=obj.get('length'), quantity=obj.get('quantity'))
                 material = Sticker(
                     type_sticker = obj.get('type_sticker'),
                     width = obj.get('width'),
                     quantity = obj.get('quantity'),
                     length = obj.get('length'),
-                    surface = surface,
+                    surface = float(obj.get('length')) * float(obj.get('width')) * float(obj.get('quantity')),
                     price_per_surface = obj.get('price_per_surface'),
                     price = obj.get('price'),
                     partiya = nakladnoy.partiya,
                     nakladnoy_id = nakladnoy.id,
-                    date=nakladnoy.date
+                    date=nakladnoy.date,
+                    provider = nakladnoy.provider
                 )
                 db.session.add(material)
             balance_minus(data.get("payed_price_d"))
@@ -294,34 +292,43 @@ def sticker_material():
     elif request.method == 'PUT' or request.method == 'PATCH':
         if user.role in ['a']:
             data = request.get_json()
-            try:
-                material_id = request.args.get("material_id")
-                material = Sticker.query.get(material_id)
-                surface = update_sticker_amount(material=material,type=data.get('type_sticker', material.type_sticker), length=data.get('length', material.length), width=data.get('width', material.width),  quantity=data.get('quantity', material.quantity))
-                material.type_sticker = data.get('type_sticker', material.type_sticker)
-                material.width = data.get('width', material.width)
-                material.length = data.get('length', material.length)
-                material.quantity = data.get('quantity', material.quantity)
-                material.surface += surface
-                db.session.commit()
-                return jsonify(msg='Success')
-            except AssertionError as err:
-                return jsonify(msg=f"{str(err)}"), 400
+            material_id = request.args.get("material_id")
+            material = Sticker.query.get(material_id)
+            if material.editable == True:
+                try:
+                    # surface = update_sticker_amount(material=material,type=data.get('type_sticker', material.type_sticker), length=data.get('length', material.length), width=data.get('width', material.width),  quantity=data.get('quantity', material.quantity))
+                    material.type_sticker = data.get('type_sticker', material.type_sticker)
+                    material.width = data.get('width', material.width)
+                    material.length = data.get('length', material.length)
+                    material.price_per_surface = data.get('price_per_surface', material.price_per_surface)
+                    material.quantity = data.get('quantity', material.quantity)
+                    material.surface = float(material.width) * float(material.length) * float(material.quantity)
+                    db.session.commit()
+                    return jsonify(msg='Success')
+                except AssertionError as err:
+                    return jsonify(msg=f"{str(err)}"), 400
+            return  jsonify(msg="Неотредактируемый объект"), 400
         return jsonify(msg="У вас нет полномочий на это действие"), 401
     elif request.method == 'DELETE':
         if user.role == 'a':
             id = request.args.get("material_id")
-            material = db.get_or_404(Sticker, id)
-            db.session.delete(material)
-            db.session.commit()
-            return jsonify(msg="Deleted")
+            nakladnoy = db.get_or_404(StickerNakladnoy, id)
+            if nakladnoy.editable == True:
+                balance_add(nakladnoy.payed_price_d)
+                stickers = Sticker.query.filter_by(nakladnoy_id=nakladnoy.id).all()
+                db.session.delete(nakladnoy)
+                for sticker in stickers:
+                    db.session.delete(sticker)
+                    db.session.commit()
+                return jsonify(msg="Deleted")
+            return  jsonify(msg="Неотредактируемый объект"), 400
         return jsonify(msg="У вас нет полномочий на это действие"), 401
 
 
 @bp.route('/makaron', methods=['GET','POST', 'DELETE'])
 def makaron():
     if request.method == 'GET':
-        makaron = Makaron.query.all()
+        makaron = Makaron.query.order_by(Makaron.date.desc()).all()
         return jsonify(makaron_schema.dump(makaron))
     elif request.method == 'POST':
         data = request.get_json()
@@ -346,7 +353,7 @@ def makaron():
 @bp.route('/granula-otxod', methods=['GET','POST', 'DELETE'])
 def granula_otxod():
     if request.method == 'GET':
-        makaron = GranulaOtxod.query.all()
+        makaron = GranulaOtxod.query.order_by(GranulaOtxod.date.desc()).all()
         return jsonify(granula_otxod_schema.dump(makaron))
     elif request.method == 'POST':
         data = request.get_json()
